Remove "dayValidator" from public API
Fixes #9

- Update docs
- Document design decisions
- Add missing semi colon, remove space
