+++ src/Popup.jsx
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 export const PopupWrapper = styled.div`
   position: relative;
   text-align: center;
-
   width: 420px;
   margin: auto;
