+++ src/calendar/MonthDays.jsx
@@ -22,7 +22,7 @@ export const MonthDay = styled.li`
   transition: all 0.25s ease;
   font-weight: ${props => (props.isToday ? 'bold' : 'inherit')};
   color: ${props => (props.isToday ? props.theme.primary : 'inherit')};
-  opacity: ${props => (props.isValid ? 1 : 0.3)}
+  opacity: ${props => (props.isValid ? 1 : 0.3)};
 
   :hover {
     cursor: ${props => (props.isValid ? 'pointer' : 'inherit')};
